requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pytz==2023.3
httpx==0.25.2
asyncpg==0.29.0
aiohttp==3.9.1
openpyxl==3.1.2
